Persist the alert messages in the handler of the middleware
Moves the persistence of the alert messages to the "handler"
method of the Alert middleware. In laravel 5.8 the sesion is
saved during the "handler" method of \Illuminate\Session\Middleware\StartSession
and not in the "terminate" method (this was the previous behaviour).

If the persistence of the messages is done in the "terminate" method,
the messages are pushed to the session but are never saved to the
persistence layer of the session. So if you call an Alert::message()
before a redirect, the messages are not saved for the next request.